backend/requirements.txt
fastapi
uvicorn
python-multipart

web3
eth-account

opencv-python
numpy

requests

python-dotenv